# pkg/api/rest/docs/swagger.yaml
basePath: /cicada
definitions:
  common.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  domain.ParameterStructure:
    properties:
      properties:
        additionalProperties:
          $ref: '#/definitions/domain.PropertyDef'
        type: object
      required:
        items:
          type: string
        type: array
    type: object
  domain.PropertyDef:
    properties:
      default: {}
      description:
        type: string
      enum:
        items:
          type: string
        type: array
      example: {}
      items:
        $ref: '#/definitions/domain.PropertyDef'
      properties:
        additionalProperties:
          $ref: '#/definitions/domain.PropertyDef'
        type: object
      required:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
  domain.TaskComponent:
    properties:
      created_at:
        type: string
      data:
        $ref: '#/definitions/domain.TaskComponentData'
      description:
        type: string
      id:
        type: string
      is_example:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    required:
    - data
    - id
    - name
    type: object
  domain.TaskComponentData:
    properties:
      body_params:
        $ref: '#/definitions/domain.ParameterStructure'
      options:
        $ref: '#/definitions/domain.TaskComponentOptions'
      path_params:
        $ref: '#/definitions/domain.ParameterStructure'
      query_params:
        $ref: '#/definitions/domain.ParameterStructure'
    type: object
  domain.TaskComponentOptions:
    properties:
      api_connection_id:
        type: string
      endpoint:
        type: string
      extra:
        additionalProperties: true
        type: object
      method:
        type: string
      request_body:
        type: string
    type: object
...
  model.CreateTaskComponentReq:
    properties:
      data:
        $ref: '#/definitions/domain.TaskComponentData'
      name:
        type: string
    required:
    - data
    - name
    type: object
...
  model.SimpleMsg:
    properties:
      message:
        type: string
    type: object
...
  model.Task:
    properties:
      dependencies:
        items:
          type: string
        type: array
      extra:
        additionalProperties: true
        type: object
      id:
        type: string
      name:
        type: string
      path_params:
        additionalProperties:
          type: string
        type: object
      query_params:
        additionalProperties:
          type: string
        type: object
      request_body:
        type: string
      task_component:
        type: string
      task_group_id:
        type: string
      workflow_id:
        type: string
    required:
    - id
    - name
    - request_body
    - task_component
    - task_group_id
    - workflow_id
    type: object
...
  model.TaskGroup:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      tasks:
        items:
          $ref: '#/definitions/model.Task'
        type: array
      workflow_id:
        type: string
    required:
    - id
    - name
    - tasks
    - workflow_id
    type: object
...
info:
  contact: {}
  description: Workflow management module
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: CM-Cicada REST API
  version: latest
paths:
...
  /task_component:
    get:
      consumes:
      - application/json
      description: Get a list of task component.
      operationId: list-task-component
      parameters:
      - description: Page of the task component list.
        in: query
        name: page
        type: string
      - description: Row of the task component list.
        in: query
        name: row
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get a list of task component.
          schema:
            items:
              $ref: '#/definitions/domain.TaskComponent'
            type: array
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get a list of task component.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: List TaskComponent
      tags:
      - '[Task Component]'
    post:
      consumes:
      - application/json
      description: Register the task component.
      operationId: create-task-component
      parameters:
      - description: task component to create.
        in: body
        name: TaskComponent
        required: true
        schema:
          $ref: '#/definitions/model.CreateTaskComponentReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully register the task component
          schema:
            $ref: '#/definitions/domain.TaskComponent'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to register the task component
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Create TaskComponent
      tags:
      - '[Task Component]'
  /task_component/{tcId}:
    delete:
      consumes:
      - application/json
      description: Delete the task component.
      operationId: delete-task-component
      parameters:
      - description: ID of the task component.
        in: path
        name: tcId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully delete the task component
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to delete the task component
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete TaskComponent
      tags:
      - '[Task Component]'
    get:
      consumes:
      - application/json
      description: Get the task component.
      operationId: get-task-component
      parameters:
      - description: ID of the TaskComponent
        in: path
        name: tcId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the task component
          schema:
            $ref: '#/definitions/domain.TaskComponent'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the task component
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get TaskComponent
      tags:
      - '[Task Component]'
    put:
      consumes:
      - application/json
      description: Update the task component.
      operationId: update-task-component
      parameters:
      - description: ID of the TaskComponent
        in: path
        name: tcId
        required: true
        type: string
      - description: task component to modify.
        in: body
        name: TaskComponent
        required: true
        schema:
          $ref: '#/definitions/model.CreateTaskComponentReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully update the task component
          schema:
            $ref: '#/definitions/domain.TaskComponent'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to update the task component
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Update TaskComponent
      tags:
      - '[Task Component]'
  /task_component/name/{tcName}:
    get:
      consumes:
      - application/json
      description: Get the task component by name.
      operationId: get-task-component-by-name
      parameters:
      - description: Name of the TaskComponent
        in: path
        name: tcName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the task component
          schema:
            $ref: '#/definitions/domain.TaskComponent'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the task component
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get TaskComponent by Name
      tags:
      - '[Task Component]'
...
  /workflow/{wfId}/task:
    get:
      consumes:
      - application/json
      description: Get a task list of the workflow.
      operationId: list-task
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get a task list.
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get a task list.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: List Task
      tags:
      - '[Workflow]'
  /workflow/{wfId}/task/{taskId}:
    get:
      consumes:
      - application/json
      description: Get the task.
      operationId: get-task
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      - description: ID of the task.
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the task.
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the task.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get Task
      tags:
      - '[Workflow]'
  /workflow/{wfId}/task_group:
    get:
      consumes:
      - application/json
      description: Get a task group list of the workflow.
      operationId: list-task-group
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get a task group list.
          schema:
            items:
              $ref: '#/definitions/model.TaskGroup'
            type: array
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get a task group list.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: List TaskGroup
      tags:
      - '[Workflow]'
  /workflow/{wfId}/task_group/{tgId}:
    get:
      consumes:
      - application/json
      description: Get the task group.
      operationId: get-task-group
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      - description: ID of the task group.
        in: path
        name: tgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the task group.
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the task group.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get TaskGroup
      tags:
      - '[Workflow]'
...
swagger: "2.0"